refactor quick date select
