+++ orval.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from 'orval';
 export default defineConfig({
   skycofl: {
     input: {
-      target: 'http://localhost:5005/api/swagger/v1/swagger.json'
+      target: 'https://sky.coflnet.com/api/swagger/v1/swagger.json'
     },
     output: {
       client: "react-query",
